m_fix : line end by pipe check
