+++ include/parse.h
@@ -29,4 +29,7 @@ int		end_by_esc(char *one_ln);
 void	skip_qmbt(char *str, int *i);
 int		check_end_esc(char *str, char *charset);
 char	**alloc_mem(char **tmp, char *start_addr, int len, int idx);
+char	*strchr_skip_bslash(const char *s, int c);
+void	is_qmbt(char *one_ln, char **qmbt);
+
 #endif
