+++ lib/libft/ft_strdup.c
@@ -21,8 +21,6 @@ char	*ft_strdup(const char *s1)
 	idx = 0;
 	len = ft_strlen(s1);
 	tmp = (char *)ft_calloc(len + 1, sizeof(char));
-	if (tmp == 0)
-		is_error(NULL, NULL, "can't allocate memory", EXIT_FAILURE);
 	while (idx < len)
 	{
 		*(tmp + idx) = *(s1 + idx);
