+++ lib/libft/ft_strjoin.c
@@ -22,8 +22,6 @@ char	*ft_strjoin(char const *s1, char const *s2)
 		return (NULL);
 	idx_1 = 0;
 	str = (char *)ft_calloc(ft_strlen(s1) + ft_strlen(s2) + 1, sizeof(char));
-	if (!str)
-		is_error(NULL, NULL, "can't allocate memory", EXIT_FAILURE);
 	while (*(s1 + idx_1))
 	{
 		*(str + idx_1) = *(s1 + idx_1);
